# inforschema: avoid mem allocs when searching table items