<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #54047

Problem Summary: avoid mem allocs when searching table items

### What changed and how does it work?

Passing `tableItem` pointer to the `matchFn` introduces mem allocs on heap. Plan cache now calls `TableByID` for each related table (ref the following flamegraph of a point-get workload), which makes `searchTableItemByID` a hot path, thus extra mem allocs should be avoid.

![image](https://github.com/user-attachments/assets/2db1bea7-06c2-4254-8a13-7a63ee049fff)

We can simply optimize it by changing the signature of `matchFn` and here is the micro-benchmark results:

```
# Baseline
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/infoschema
cpu: Intel(R) Xeon(R) Gold 5220R CPU @ 2.20GHz
BenchmarkSearchByID_100
BenchmarkSearchByID_100-32               1898032               648.4 ns/op           288 B/op          3 allocs/op
BenchmarkSearchByID_1000
BenchmarkSearchByID_1000-32              1763356               714.4 ns/op           288 B/op          3 allocs/op
BenchmarkSearchByID_10000
BenchmarkSearchByID_10000-32             1657314               756.6 ns/op           288 B/op          3 allocs/op

# This PR
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/infoschema
cpu: Intel(R) Xeon(R) Gold 5220R CPU @ 2.20GHz
BenchmarkSearchByID_100
BenchmarkSearchByID_100-44               3701288               297.9 ns/op             0 B/op          0 allocs/op
BenchmarkSearchByID_1000
BenchmarkSearchByID_1000-44              3440937               306.4 ns/op             0 B/op          0 allocs/op
BenchmarkSearchByID_10000
BenchmarkSearchByID_10000-44             3548060               423.7 ns/op             0 B/op          0 allocs/op
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```